requirements.txt
python-dateutil>=2.8.2
tabulate>=0.9.0
Pillow>=9.0.0
python-xlib>=0.31
duckdb>=0.6.1

